Refactor into separate packages for view and model
