+++ app/src/main/res/layout/activity_rich_text_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RichTextViewActivity">
+    tools:context=".view.RichTextViewActivity">
 
     <WebView
         android:id="@+id/webview"
